+++ alacritty/src/config/mod.rs
@@ -202,6 +202,15 @@ fn print_deprecation_warnings(config: &Config) {
             "Config persistent_logging is deprecated; please use debug.persistent_logging instead"
         );
     }
+
+    if config.scrolling.faux_multiplier().is_some() {
+        warn!(
+            target: LOG_TARGET_CONFIG,
+            "Config scrolling.faux_multiplier is deprecated; the alternate scroll escape can now \
+             be used to disable it and `scrolling.multiplier` controls the number of scrolled \
+             lines"
+        );
+    }
 }
 
 #[cfg(test)]